Bug Fix : copy_to_user and '\0'
In the previous version of program, copy_to_user would only copy
strlen numbers of character to the buffer, which would miss the
last '\0' character.

This would not pose a problem when we find Fibonacci number
increasingly, however, when we find the number decreasingly, as we
do not give the ending point of the character array, the program
would continue look for the end point which is from the previous
iteration.